refactor: extract the break screen layout into a shared intervention screen
