+++ app/src/main/java/com/flx_apps/digitaldetox/ui/screens/feature/break_doom_scrolling/BreakDoomScrollingOverlay.kt
@@ -1,62 +1,26 @@
 package com.flx_apps.digitaldetox.ui.screens.feature.break_doom_scrolling
 
 import android.content.Intent
-import androidx.compose.animation.core.Animatable
-import androidx.compose.animation.core.FastOutSlowInEasing
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.SelfImprovement
 import androidx.compose.material.icons.filled.WavingHand
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.rememberVectorPainter
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import com.flx_apps.digitaldetox.R
 import com.flx_apps.digitaldetox.features.BreakDoomScrollingFeature
 import com.flx_apps.digitaldetox.system_integration.OverlayContent
 import com.flx_apps.digitaldetox.system_integration.OverlayService
-import com.flx_apps.digitaldetox.ui.theme.DetoxDroidTheme
+import com.flx_apps.digitaldetox.ui.widgets.DrainingBar
+import com.flx_apps.digitaldetox.ui.widgets.InterventionPrimaryButton
+import com.flx_apps.digitaldetox.ui.widgets.InterventionScreen
+import com.flx_apps.digitaldetox.ui.widgets.InterventionSecondaryButton
 import com.flx_apps.digitaldetox.util.ForceStopUtil
 
 /**
@@ -152,173 +116,44 @@ fun BreakDoomScrollingOverlayContent(
     onExitApp: () -> Unit,
     onFinishFirst: () -> Unit,
 ) {
-    DetoxDroidTheme {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(
-                    Brush.verticalGradient(
-                        listOf(Color(0xF5141420), Color(0xFB08080D), Color(0xFF000000))
-                    )
-                )
-                .padding(horizontal = 32.dp), horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Spacer(modifier = Modifier.weight(1.2f))
-            BreathingBadge(mode)
-            Text(
-                text = stringResource(
-                    id = when (mode) {
-                        BreakScreenMode.WARNING -> R.string.feature_doomScrolling_warning_title
-                        BreakScreenMode.COOLDOWN -> R.string.feature_doomScrolling_cooldown_title
-                        BreakScreenMode.GUIDE_OUT -> R.string.feature_doomScrolling_guideOut_title
-                    }
-                ),
-                style = MaterialTheme.typography.displaySmall,
-                textAlign = TextAlign.Center,
-                color = Color.White,
-                modifier = Modifier.padding(top = 32.dp)
-            )
-            Text(
-                text = stringResource(
-                    id = when (mode) {
-                        BreakScreenMode.WARNING -> R.string.feature_doomScrolling_warning_message
-                        BreakScreenMode.COOLDOWN -> R.string.feature_doomScrolling_cooldown_message
-                        BreakScreenMode.GUIDE_OUT -> R.string.feature_doomScrolling_guideOut_message
-                    }
-                ),
-                style = MaterialTheme.typography.titleMedium,
-                textAlign = TextAlign.Center,
-                color = Color.White.copy(alpha = 0.85f),
-                modifier = Modifier.padding(top = 16.dp)
-            )
-            contextText?.let {
-                Text(
-                    text = it,
-                    style = MaterialTheme.typography.bodyMedium,
-                    textAlign = TextAlign.Center,
-                    color = Color.White.copy(alpha = 0.75f),
-                    modifier = Modifier
-                        .padding(top = 24.dp)
-                        .clip(RoundedCornerShape(20.dp))
-                        .background(Color.White.copy(alpha = 0.08f))
-                        .padding(horizontal = 20.dp, vertical = 10.dp)
-                )
+    InterventionScreen(
+        icon = when (mode) {
+            BreakScreenMode.WARNING -> painterResource(id = R.drawable.ic_scroll)
+            BreakScreenMode.COOLDOWN -> rememberVectorPainter(Icons.Default.SelfImprovement)
+            BreakScreenMode.GUIDE_OUT -> rememberVectorPainter(Icons.Default.WavingHand)
+        },
+        title = stringResource(
+            id = when (mode) {
+                BreakScreenMode.WARNING -> R.string.feature_doomScrolling_warning_title
+                BreakScreenMode.COOLDOWN -> R.string.feature_doomScrolling_cooldown_title
+                BreakScreenMode.GUIDE_OUT -> R.string.feature_doomScrolling_guideOut_title
             }
-            if (mode == BreakScreenMode.GUIDE_OUT) {
-                GuideOutCountdown(onFinished = onExitApp)
-            }
-            Button(
-                onClick = onExitApp,
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color.White, contentColor = Color.Black
-                ),
-                contentPadding = PaddingValues(horizontal = 36.dp, vertical = 14.dp),
-                modifier = Modifier.padding(top = 32.dp)
-            ) {
-                Text(
-                    text = stringResource(id = R.string.feature_doomScrolling_warning_exit),
-                    style = MaterialTheme.typography.titleMedium
-                )
-            }
-            if (mode == BreakScreenMode.WARNING) {
-                TextButton(
-                    modifier = Modifier.padding(top = 8.dp), onClick = onFinishFirst
-                ) {
-                    Text(
-                        text = stringResource(id = R.string.feature_doomScrolling_warning_finishFirst),
-                        color = Color.White.copy(alpha = 0.7f)
-                    )
-                }
-            }
-            Spacer(modifier = Modifier.weight(1f))
-            Image(
-                painter = painterResource(id = R.drawable.ic_launcher_foreground_cropped),
-                contentDescription = null,
-                // the drawable carries a 15% transparent safe-zone margin on every side (its
-                // artwork is wrapped in a scale(0.7) group); shifting it down by exactly that
-                // margin puts the droid flush on the bottom screen edge
-                modifier = Modifier
-                    .size(LOGO_SIZE)
-                    .offset(y = LOGO_SIZE * 0.15f)
-            )
-        }
-    }
-}
-
-/**
- * A slowly pulsing ("breathing") circular badge with a mode-specific icon — a calm visual anchor
- * that sets the pace against the frantic scrolling that led here.
- */
-@Composable
-private fun BreathingBadge(mode: BreakScreenMode) {
-    val breath = rememberInfiniteTransition(label = "breathing")
-    val scale by breath.animateFloat(
-        initialValue = 1f,
-        targetValue = 1.08f,
-        animationSpec = infiniteRepeatable(
-            tween(durationMillis = 2400, easing = FastOutSlowInEasing), RepeatMode.Reverse
         ),
-        label = "breathingScale"
-    )
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier
-            .size(104.dp)
-            .scale(scale)
-            .clip(CircleShape)
-            .background(Color.White.copy(alpha = 0.06f))
-            .border(1.dp, Color.White.copy(alpha = 0.15f), CircleShape)
+        message = stringResource(
+            id = when (mode) {
+                BreakScreenMode.WARNING -> R.string.feature_doomScrolling_warning_message
+                BreakScreenMode.COOLDOWN -> R.string.feature_doomScrolling_cooldown_message
+                BreakScreenMode.GUIDE_OUT -> R.string.feature_doomScrolling_guideOut_message
+            }
+        ),
+        contextText = contextText,
     ) {
-        val tint = Color.White.copy(alpha = 0.9f)
-        val iconModifier = Modifier.size(44.dp)
-        when (mode) {
-            BreakScreenMode.WARNING -> Icon(
-                painterResource(id = R.drawable.ic_scroll), null, iconModifier, tint
-            )
-
-            BreakScreenMode.COOLDOWN -> Icon(
-                Icons.Default.SelfImprovement, null, iconModifier, tint
-            )
-
-            BreakScreenMode.GUIDE_OUT -> Icon(
-                Icons.Default.WavingHand, null, iconModifier, tint
+        if (mode == BreakScreenMode.GUIDE_OUT) {
+            DrainingBar(
+                durationMs = BreakDoomScrollingOverlayService.GUIDE_OUT_AUTO_EXIT_MS,
+                onFinished = onExitApp
             )
         }
-    }
-}
-
-/**
- * A thin bar draining from full to empty over
- * [BreakDoomScrollingOverlayService.GUIDE_OUT_AUTO_EXIT_MS]; calls [onFinished] when it runs out.
- * It owns the guide-out timing, so the countdown and the actual exit can never drift apart.
- */
-@Composable
-private fun GuideOutCountdown(onFinished: () -> Unit) {
-    val progress = remember { Animatable(1f) }
-    LaunchedEffect(Unit) {
-        progress.animateTo(
-            targetValue = 0f,
-            animationSpec = tween(
-                durationMillis = BreakDoomScrollingOverlayService.GUIDE_OUT_AUTO_EXIT_MS.toInt(),
-                easing = LinearEasing
+        InterventionPrimaryButton(
+            text = stringResource(id = R.string.feature_doomScrolling_warning_exit),
+            onClick = onExitApp
+        )
+        if (mode == BreakScreenMode.WARNING) {
+            InterventionSecondaryButton(
+                text = stringResource(id = R.string.feature_doomScrolling_warning_finishFirst),
+                onClick = onFinishFirst
             )
-        )
-        onFinished()
-    }
-    Box(
-        modifier = Modifier
-            .padding(top = 24.dp)
-            .width(160.dp)
-            .height(4.dp)
-            .clip(RoundedCornerShape(2.dp))
-            .background(Color.White.copy(alpha = 0.2f))
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth(progress.value)
-                .height(4.dp)
-                .background(Color.White.copy(alpha = 0.85f))
-        )
+        }
     }
 }
 
@@ -354,5 +189,3 @@ private fun BreakDoomScrollingGuideOutPreview() {
         onFinishFirst = {},
     )
 }
-
-private val LOGO_SIZE = 196.dp
